% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eventdrop.R
\name{eventdrop}
\alias{eventdrop}
\title{Create an EventDrop timeline}
\usage{
eventdrop(data = NULL, name = NULL, date = NULL, ..., width = NULL,
  height = NULL, elementId = NULL)
}
\description{
eventdrop provides an interactive \code{d3.js} htmlwidget
for \href{https://github.com/marmelab/EventDrop}{EventDrop}
timelines.
}

