% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eventdrop.R
\name{eventdrop}
\alias{eventdrop}
\title{Create an EventDrop timeline}
\usage{
eventdrop(data = NULL, name = NULL, date = NULL, ..., width = NULL,
  height = NULL, elementId = NULL)
}
\description{
<Add Description>
}

